refactor: LineParser 클래스들을 기존 클래스들과 패키지 분리
